@@ -5,10 +5,10 @@
     "type": "library",
     "license": "BSD-3-Clause",
     "require": {
-        "php": "^7.0",
+        "php": "^7.4",
         "psr/http-message": "^1.0",
         "linio/microlog": "^1.0",
-        "linio/util": "^2.1",
+        "linio/util": "^3.0",
         "particle/filter": "^1.4",
         "particle/validator": "^2.1",
         "monolog/monolog": "^1.20",
@@ -16,9 +16,11 @@
         "zendframework/zend-expressive-router": "^1.1"
     },
     "require-dev": {
-        "phpunit/phpunit": "^5.4 || ^7.0",
-        "friendsofphp/php-cs-fixer": "^2.4",
-        "eloquent/phony-phpunit": "^1.0 || ^4.0"
+        "ext-json": "*",
+        "phpunit/phpunit": "^9.5",
+        "phpspec/prophecy-phpunit": "^2.0",
+        "friendsofphp/php-cs-fixer": "^3.0",
+        "eloquent/phony-phpunit": "^7.1"
     },
     "autoload": {
         "files": [
@@ -30,7 +32,8 @@
     },
     "autoload-dev": {
         "psr-4": {
-            "Linio\\TestAssets\\": "tests/assets"
+            "Linio\\TestAssets\\": "tests/assets",
+            "Linio\\Common\\Expressive\\Tests\\": "tests/"
         }
     }
 }
