@@ -8,32 +8,15 @@ use Particle\Filter\Filter;
 
 class FilterService
 {
-    /**
-     * @var string
-     */
-    private $filterClass;
+    private string $filterClass;
+    private FilterRulesFactory $filterRulesFactory;
 
-    /**
-     * @var FilterRulesFactory
-     */
-    private $filterRulesFactory;
-
-    /**
-     * @param string $filterClass
-     * @param FilterRulesFactory $filterRulesFactory
-     */
     public function __construct(string $filterClass, FilterRulesFactory $filterRulesFactory)
     {
         $this->filterClass = $filterClass;
         $this->filterRulesFactory = $filterRulesFactory;
     }
 
-    /**
-     * @param array $input
-     * @param array $filterRulesClasses
-     *
-     * @return array
-     */
     public function filter(array $input, array $filterRulesClasses): array
     {
         if (empty($filterRulesClasses)) {
