@@ -11,26 +11,13 @@ use Zend\Expressive\Router\RouteResult;
 
 class ConfigureNewrelicForRequest
 {
-    /**
-     * @var string
-     */
-    private $appName;
+    private string $appName;
 
-    /**
-     * @param string $appName
-     */
     public function __construct(string $appName)
     {
         $this->appName = $appName;
     }
 
-    /**
-     * @param ServerRequestInterface $request
-     * @param ResponseInterface $response
-     * @param callable $next
-     *
-     * @return ResponseInterface
-     */
     public function __invoke(ServerRequestInterface $request, ResponseInterface $response, callable $next): ResponseInterface
     {
         if (!extension_loaded('newrelic')) {
@@ -44,9 +31,6 @@ class ConfigureNewrelicForRequest
         return $next($request, $response);
     }
 
-    /**
-     * @param ServerRequestInterface $request
-     */
     private function nameRouteIfRouteFound(ServerRequestInterface $request)
     {
         /** @var RouteResult $routeResult */
@@ -62,8 +46,6 @@ class ConfigureNewrelicForRequest
     }
 
     /**
-     * @param ServerRequestInterface $request
-     *
      * @throws MiddlewareOutOfOrderException
      */
     private function addRequestIdToNewrelic(ServerRequestInterface $request)
