@@ -10,20 +10,9 @@ use Particle\Validator\Validator;
 
 class ValidationService
 {
-    /**
-     * @var ValidatorFactory
-     */
-    private $validatorFactory;
+    private ValidatorFactory $validatorFactory;
+    private ValidationRulesFactory $validationRulesFactory;
 
-    /**
-     * @var ValidationRulesFactory
-     */
-    private $validationRulesFactory;
-
-    /**
-     * @param ValidatorFactory $validatorFactory
-     * @param ValidationRulesFactory $validationRulesFactory
-     */
     public function __construct(ValidatorFactory $validatorFactory, ValidationRulesFactory $validationRulesFactory)
     {
         $this->validatorFactory = $validatorFactory;
@@ -31,9 +20,6 @@ class ValidationService
     }
 
     /**
-     * @param array $input
-     * @param array $validationRulesClasses
-     *
      * @throws NotFoundException
      * @throws InvalidRequestException
      */
@@ -59,8 +45,6 @@ class ValidationService
 
     /**
      * Maps the validator's errors to DomainException's error format.
-     *
-     * @param array $errors
      *
      * @throws InvalidRequestException
      */
