@@ -2,56 +2,70 @@
 
 declare(strict_types=1);
 
-namespace Linio\Common\Expressive\Filter;
+namespace Linio\Common\Expressive\Tests\Filter;
 
-use Eloquent\Phony\Phpunit\Phony;
 use Interop\Container\ContainerInterface;
 use Linio\Common\Expressive\Exception\Base\NotFoundException;
+use Linio\Common\Expressive\Filter\FilterRulesFactory;
 use Linio\TestAssets\TestFilterRules;
 use PHPUnit\Framework\TestCase;
+use Prophecy\PhpUnit\ProphecyTrait;
 
 class FilterRulesFactoryTest extends TestCase
 {
+    use ProphecyTrait;
+
     public function testItGetsTheFilterRulesFromTheContainer()
     {
         $class = TestFilterRules::class;
         $testFilterRules = new $class();
 
-        $container = Phony::mock(ContainerInterface::class);
-        $container->has->with($class)->returns(true);
-        $container->get->with($class)->returns($testFilterRules);
+        $container = $this->prophesize(ContainerInterface::class);
+        $container
+            ->has($class)
+            ->shouldBeCalled()
+            ->willReturn(true);
+        $container
+            ->get($class)
+            ->shouldBeCalled()
+            ->willReturn($testFilterRules);
 
-        $factory = new FilterRulesFactory($container->get());
+        $factory = new FilterRulesFactory($container->reveal());
         $factory->make($class);
-
-        $container->has->calledWith($class);
-        $container->get->calledWith($class);
     }
 
     public function testItInstantiatesTheFilterRulesWhenItIsntInTheContainer()
     {
         $class = TestFilterRules::class;
 
-        $container = Phony::mock(ContainerInterface::class);
-        $container->has->with($class)->returns(false);
+        $container = $this->prophesize(ContainerInterface::class);
+        $container
+            ->has($class)
+            ->shouldBeCalled()
+            ->willReturn(false);
+        $container
+            ->get($class)
+            ->shouldNotBeCalled();
 
-        $factory = new FilterRulesFactory($container->get());
+        $factory = new FilterRulesFactory($container->reveal());
         $actual = $factory->make($class);
 
         $this->assertInstanceOf($class, $actual);
-        $container->get->never()->calledWith($class);
     }
 
     public function testItFailsIfTheFilterRulesClassDoesntExist()
     {
         $class = 'InvalidClass';
 
-        $container = Phony::mock(ContainerInterface::class);
-        $container->has->with($class)->returns(false);
+        $container = $this->prophesize(ContainerInterface::class);
+        $container
+            ->has($class)
+            ->shouldBeCalled()
+            ->willReturn(false);
 
         $this->expectException(NotFoundException::class);
 
-        $factory = new FilterRulesFactory($container->get());
+        $factory = new FilterRulesFactory($container->reveal());
         $factory->make($class);
     }
 }
