@@ -2,9 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Linio\Common\Expressive\Middleware;
+namespace Linio\Common\Expressive\Tests\Middleware;
 
 use Linio\Common\Expressive\Exception\Http\MiddlewareOutOfOrderException;
+use Linio\Common\Expressive\Middleware\AddRequestIdToLog;
 use PHPUnit\Framework\TestCase;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
