@@ -2,10 +2,14 @@
 
 declare(strict_types=1);
 
-namespace Linio\Common\Expressive\Middleware;
+namespace Linio\Common\Expressive\Tests\Middleware;
 
 use Linio\Common\Expressive\Exception\Http\MiddlewareOutOfOrderException;
+use Linio\Common\Expressive\Middleware\AddRequestIdToLog;
+use Linio\Common\Expressive\Middleware\AddRequestIdToResponse;
 use PHPUnit\Framework\TestCase;
+use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\ServerRequestInterface;
 use Zend\Diactoros\Response;
 use Zend\Diactoros\Response\EmptyResponse;
 use Zend\Diactoros\ServerRequest;
@@ -18,7 +22,7 @@ class AddRequestIdToResponseTest extends TestCase
 
         $request = (new ServerRequest())->withAttribute('requestId', $requestId);
         $response = new Response();
-        $callable = function ($request, $response) use ($requestId) {
+        $callable = function ($request, $response) {
             return new EmptyResponse();
         };
 
