@@ -2,14 +2,15 @@
 
 declare(strict_types=1);
 
-namespace Linio\Common\Expressive\Middleware;
+namespace Linio\Common\Expressive\Tests\Middleware;
 
-use Eloquent\Phony\Phpunit\Phony;
 use Linio\Common\Expressive\Exception\Http\MiddlewareOutOfOrderException;
 use Linio\Common\Expressive\Exception\Http\RouteNotFoundException;
+use Linio\Common\Expressive\Middleware\ValidateRequestBody;
 use Linio\Common\Expressive\Validation\ValidationService;
 use Linio\TestAssets\TestValidationRules;
 use PHPUnit\Framework\TestCase;
+use Prophecy\PhpUnit\ProphecyTrait;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Zend\Diactoros\Response;
@@ -19,9 +20,11 @@ use Zend\Expressive\Router\RouteResult;
 
 class ValidateRequestBodyTest extends TestCase
 {
+    use ProphecyTrait;
+
     public function testItSkipsValidationIfTheRouterHasntRun()
     {
-        $validationService = Phony::mock(ValidationService::class);
+        $validationService = $this->prophesize(ValidationService::class);
 
         $request = new ServerRequest();
         $response = new Response();
@@ -31,13 +34,13 @@ class ValidateRequestBodyTest extends TestCase
 
         $this->expectException(MiddlewareOutOfOrderException::class);
 
-        $middleware = new ValidateRequestBody($validationService->get(), []);
+        $middleware = new ValidateRequestBody($validationService->reveal(), []);
         $middleware->__invoke($request, $response, $next);
     }
 
     public function testItSkipsValidationIfARouteIsNotFound()
     {
-        $validationService = Phony::mock(ValidationService::class);
+        $validationService = $this->prophesize(ValidationService::class);
 
         $routeResult = RouteResult::fromRouteFailure();
         $request = (new ServerRequest())->withAttribute(RouteResult::class, $routeResult);
@@ -49,15 +52,13 @@ class ValidateRequestBodyTest extends TestCase
 
         $this->expectException(MiddlewareOutOfOrderException::class);
 
-        $middleware = new ValidateRequestBody($validationService->get(), []);
+        $middleware = new ValidateRequestBody($validationService->reveal(), []);
         $middleware->__invoke($request, $response, $next);
     }
 
     public function testItFailsValidationIfTheRouteIsNotFoundInRoutes()
     {
-        $routes = require __DIR__ . '/../assets/routes.php';
-
-        $validationService = Phony::mock(ValidationService::class);
+        $validationService = $this->prophesize(ValidationService::class);
 
         $routeResult = RouteResult::fromRouteMatch('invalid', 'TestMiddleware', []);
         $request = (new ServerRequest())->withAttribute(RouteResult::class, $routeResult);
@@ -69,7 +70,7 @@ class ValidateRequestBodyTest extends TestCase
 
         $this->expectException(RouteNotFoundException::class);
 
-        $middleware = new ValidateRequestBody($validationService->get(), []);
+        $middleware = new ValidateRequestBody($validationService->reveal(), []);
         $middleware->__invoke($request, $response, $next);
     }
 
@@ -77,7 +78,10 @@ class ValidateRequestBodyTest extends TestCase
     {
         $routes = require __DIR__ . '/../assets/routes.php';
 
-        $validationService = Phony::mock(ValidationService::class);
+        $validationService = $this->prophesize(ValidationService::class);
+        $validationService
+            ->validate([], [TestValidationRules::class])
+            ->shouldBeCalled();
 
         $routeResult = RouteResult::fromRouteMatch('test', 'TestMiddleware', []);
         $request = (new ServerRequest())->withAttribute(RouteResult::class, $routeResult)->withParsedBody([]);
@@ -87,9 +91,7 @@ class ValidateRequestBodyTest extends TestCase
             return new EmptyResponse();
         };
 
-        $middleware = new ValidateRequestBody($validationService->get(), $routes);
+        $middleware = new ValidateRequestBody($validationService->reveal(), $routes);
         $middleware->__invoke($request, $response, $next);
-
-        $validationService->validate->calledWith([], [TestValidationRules::class]);
     }
 }
